セキュリティのためフォルダ回りの処理を変更しました。
git-svn-id: http://aipo.googlecode.com/svn/trunk@925 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ portlets/webmail/src/main/java/com/aimluck/eip/webmail/WebMailFolderSelectData.java
@@ -163,7 +163,7 @@ public class WebMailFolderSelectData extends
 
       WebMailFolderResultData rd = new WebMailFolderResultData();
       rd.initField();
-      rd.setFolderId(record.getFolderId().longValue());
+      rd.setFolderId(record.getFolderId());
       rd.setFolderName(record.getFolderName());
       rd.setCreateDate(ALDateUtil.format(record.getCreateDate(), "yyyy年M月d日"));
